requirements.txt
Flask==3.0.2
matplotlib==3.7.5
numpy==1.24.3
onnxruntime==1.17.1
opencv-python-headless==4.5.3.56  # Headless version of OpenCV, if GUI is not required
Pillow==10.3.0
pycocotools==2.0.7
Requests==2.31.0
setuptools==68.2.2
torch==2.0.1
torchvision==0.15.2
requests-toolbelt==1.0.0